requirements.txt
python-telegram-bot
requests
image
winshell
pypiwin32
pyinstaller
psutil
pillow